lab-project/requirements.txt
dotenv
azure-ai-inference
